pyproject.toml
[project]
name = "ai-competitor-research-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic==1.13.1",
    "apscheduler==3.10.4",
    "dagshub>=0.6.0",
    "fastapi==0.111.0",
    "feedparser==6.0.11",
    "groq==0.9.0",
    "httpx==0.27.0",
    "mlflow>=3.1.0",
    "pandas==2.2.2",
    "playwright==1.44.0",
    "plotly==5.22.0",
    "prometheus-fastapi-instrumentator==7.0.0",
    "pydantic==2.7.1",
    "pydantic-settings==2.3.1",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.1",
    "requests==2.32.3",
    "sqlalchemy==2.0.30",
    "streamlit==1.35.0",
    "uvicorn[standard]==0.30.1",
]
